Remake paste_timestamp logic + add pasted timestamp to history
This should work with any time format no matter how unintuitive it might
be:

01:01:01.123
00:61:00.123
01:61:61.123
61:61
00:59:61
600.123
etc…

Every timestamp is valid unless it's only milliseconds (.123) or the
timestamp exceeds the duration of video.

Additionally it now adds the pasted timestamp to history list and turns
itself off after valid timestamp was pasted.