Timestamp shifting fix + Repurpose timer for a better selection indicator
- By default seek-to uses a bold tag to indicate selected digit, which
doesn't work when user uses `osd-bold=yes`, this repurposes the redraw
timer for an underline indicator with a changed color `selection_color`
of selected digit.

- There was a width issue when pairs of 1's were beside each other
(11:11:11.111), making the timestamp visually shift when moving
selection left/right, this was 'fixed' by adding a `thin space` and
forcing its scale to 0. On top of that bold font is forced for the whole
message independently of users `osd-bold` setting to even more minimise
the visual "empty space" caused by thin space. For any reason it can
still be turned off by removing bold tag: `fb` -> `fb = ""`

This was only tested on mpv's default font on win10, i'm not certain how
it behaves on different systems or when user himself defines a different
osd font.

- Font border color for selected digit can be added in
`selection_border_color`